Revert "i8 again one thruster"
This reverts commit fc89f2b554557d24346231ffe62233d370945793.

+++ src/hyke_vessel_msgs/hyke_vessel_control_msgs.rs
@@ -3,7 +3,7 @@ use serde_derive::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, PartialEq, Clone, Copy)]
 pub struct HykeThrusterCommand {
-    pub i: i8, // index
+    pub i: i32, // index
     pub u: f32, // power / force
     pub alpha: f32, // angle
 }
